requirements.txt
fastapi==0.111.0
SQLAlchemy==1.4.0
psycopg2-binary
alembic==1.13.1
PyJWT==2.8.0
passlib==1.7.4
python-decouple==3.8
bcrypt==4.0.1
requests==2.32.3